Close oppgave where behandler corrected their response
Try this again with correct uuid, last commit was wrong.
2f221d953fe999b721e8c849d69ada3e3d37cda4 used oppgave uuid instead of referanse_uuid.
This oppgave is not possible to complete because behandler changed their response from 'NYTT_TID_STED' to 'KOMMER'.
This causes us to create an oppgave for the first response, but then hide the "behandle" button because the latest answer is 'KOMMER'.
https://jira.adeo.no/browse/FAGSYSTEM-253198